replace  file and folder for AddressSpaceFilter and AddressSpaceToolbar
